+++ server/app.js
@@ -7,6 +7,8 @@ var logger = require('morgan');
 var indexRouter = require('./routes/index');
 var productRouter = require('./routes/product');
 var categoryRouter = require('./routes/category');
+var reportRouter = require('./routes/report');
+var supplierRouter = require('./routes/supplier');
 
 var app = express();
 
@@ -29,6 +31,8 @@ app.use(function(req, res, next) {
 app.use('/', indexRouter);
 app.use('/products', productRouter);
 app.use('/category', categoryRouter);
+app.use('/report', reportRouter);
+app.use('/supplier', supplierRouter);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {