+++ webclient/src/components/purchase/PurchaseList.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function PurchaseFormList() {
+    return <div>purchase List</div>
+}