+++ webclient/src/components/purchase/PurchaseView.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function PurchaseView() {
+    return <div>purchase view</div>
+}